LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/self-et-al.window-expected.txt


FAIL iframeWindow.frames before and after removal assert_equals: frames got cleared after browsing context removal expected object "[object Window]" but got null
FAIL popupWindow.frames before, after closing, and after discarding assert_true: expected true got false
PASS iframeWindow.globalThis before and after removal 
FAIL popupWindow.globalThis before, after closing, and after discarding assert_true: expected true got false
FAIL iframeWindow.self before and after removal assert_equals: self got cleared after browsing context removal expected object "[object Window]" but got null
FAIL popupWindow.self before, after closing, and after discarding assert_true: expected true got false
FAIL iframeWindow.window before and after removal assert_equals: window got cleared after browsing context removal expected object "[object Window]" but got null
FAIL popupWindow.window before, after closing, and after discarding assert_true: expected true got false
